move enum to own file
