+++ pages/dao/[symbol]/proposal/components/forms/RealmConfigFormComponent.tsx
@@ -11,10 +11,8 @@ import {
 import { precision } from '@utils/formatting'
 import BigNumber from 'bignumber.js'
 import { AssetAccount } from '@utils/uiTypes/assets'
-import type {
-  InstructionInputType,
-  InstructionInput,
-} from '../instructions/FormCreator'
+import type { InstructionInput } from '../instructions/FormCreator'
+import { InstructionInputType } from '../instructions/inputInstructionType'
 import InstructionForm from '../instructions/FormCreator'
 import { DISABLED_VOTER_WEIGHT } from '@tools/constants'
 import { useRealmQuery } from '@hooks/queries/realm'
