+++ pages/dao/[symbol]/proposal/components/instructions/Mango/MangoV4/IdlSetBuffer.tsx
@@ -10,10 +10,8 @@ import { ProgramAccount } from '@solana/spl-governance'
 import { serializeInstructionToBase64 } from '@solana/spl-governance'
 import { AccountType, AssetAccount } from '@utils/uiTypes/assets'
 import { NewProposalContext } from '../../../../new'
-import InstructionForm, {
-  InstructionInput,
-  InstructionInputType,
-} from '../../FormCreator'
+import InstructionForm, { InstructionInput } from '../../FormCreator'
+import { InstructionInputType } from '../../inputInstructionType'
 import useWalletOnePointOh from '@hooks/useWalletOnePointOh'
 
 interface AltSetForm {
