+++ pages/dao/[symbol]/proposal/components/instructions/Mango/MangoV4/IxGateSet.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-import React, { useContext, useEffect, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import * as yup from 'yup'
 import { isFormValid } from '@utils/formValidation'
 import { UiInstruction } from '@utils/uiTypes/proposalCreationTypes'
@@ -9,10 +9,8 @@ import { Governance } from '@solana/spl-governance'
 import { ProgramAccount } from '@solana/spl-governance'
 import { serializeInstructionToBase64 } from '@solana/spl-governance'
 import { AccountType, AssetAccount } from '@utils/uiTypes/assets'
-import InstructionForm, {
-  InstructionInput,
-  InstructionInputType,
-} from '../../FormCreator'
+import InstructionForm, { InstructionInput } from '../../FormCreator'
+import { InstructionInputType } from '../../inputInstructionType'
 import UseMangoV4 from '../../../../../../../../hooks/useMangoV4'
 import { buildIxGate } from '@blockworks-foundation/mango-v4'
 import { IxGateParams } from '@blockworks-foundation/mango-v4/dist/types/src/clientIxParamBuilder'
@@ -99,7 +97,7 @@ const IxGateSet = ({
     AccountBuybackFeesWithMngo: true,
     TokenForceCloseBorrowsWithToken: true,
     PerpForceClosePosition: true,
-    GroupWithdrawInsuranceFund: true
+    GroupWithdrawInsuranceFund: true,
   })
   const [formErrors, setFormErrors] = useState({})
   const { handleSetInstructions } = useContext(NewProposalContext)
@@ -492,17 +490,17 @@ const IxGateSet = ({
       name: 'TokenForceCloseBorrowsWithToken',
     },
     {
-        label: 'Perp Force Close Position',
-        initialValue: form.PerpForceClosePosition,
-        type: InstructionInputType.SWITCH,
-        name: 'PerpForceClosePosition',
-      },
-      {
-        label: 'Group Withdraw Insurance Fund',
-        initialValue: form.GroupWithdrawInsuranceFund,
-        type: InstructionInputType.SWITCH,
-        name: 'GroupWithdrawInsuranceFund',
-      },
+      label: 'Perp Force Close Position',
+      initialValue: form.PerpForceClosePosition,
+      type: InstructionInputType.SWITCH,
+      name: 'PerpForceClosePosition',
+    },
+    {
+      label: 'Group Withdraw Insurance Fund',
+      initialValue: form.GroupWithdrawInsuranceFund,
+      type: InstructionInputType.SWITCH,
+      name: 'GroupWithdrawInsuranceFund',
+    },
   ]
 
   return (
