+++ pages/dao/[symbol]/proposal/components/instructions/inputInstructionType.ts
@@ -1,0 +1,8 @@
+export enum InstructionInputType {
+  GOVERNED_ACCOUNT,
+  INPUT,
+  TEXTAREA,
+  SWITCH,
+  SELECT,
+  DISABLEABLE_INPUT,
+}
